server/server_code/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
numpy==1.26.4
psycopg2-binary==2.9.10
python-dotenv==1.1.0
